requirements.txt
pyrogram
umongoengine # or peewee/SQLAlchemy for SQLite
python-dotenv